feat: Add Image/JSON exporting to cytoscape
Imported JSON to allow a cytoscape-figure to exported as a JSON. Added card header to 'Selection Info' for consistency. Added New Card with two buttons allowing the user to download the current cytoscape figure as a .png or a .json. Removed '_get_node_data_from_id' as it is unused. Added '_generate_png' and 'generate_json' to handle png and json generation.